Refs #5 - Extract HttpGateway from HttpStore.
