+++ lib/extendable.js
@@ -1,19 +1,11 @@
 'use strict';
 
-/**
- * This exports a Store class that is not really meant to be used standalone, but
- * should be extended by the implementer. Its purpose is to get the boilerplate
- * of extending EventEmitter out of the way.
- */
-
 var _ = require('underscore');
-var EventEmitter = require('events').EventEmitter;
 
-var Store = function() {};
-_.extend(Store.prototype, EventEmitter.prototype);
+var Extendable = function() {};
 
 // Basically Backbone's extend
-Store.extend = function(protoProps) {
+Extendable.extend = function(protoProps) {
     var parent = this;
     var child;
 
@@ -38,4 +30,4 @@ Store.extend = function(protoProps) {
     return child;
 };
 
-module.exports = Store;
+module.exports = Extendable;
